% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AoristicFunctions.R
\name{aoristic_df}
\alias{aoristic_df}
\title{Create an aoristic data frame.}
\usage{
aoristic_df(start, end, rest = NULL)
}
\arguments{
\item{start}{Start of the interval.}

\item{end}{End of the interval.}

\item{rest}{Data frame; All other columns to add to the aoristic data frame.}
}
\value{
A data frame of class \code{"aoristic_df"} with at least columns
  \code{t_start} and \code{t_end},
}
\description{
Create an aoristic data frame.
}
