% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{aoristic_vm_mle}
\alias{aoristic_vm_mle}
\title{Get maximum likelihood estimates for the von Mises distribution based on
aoristic data.}
\usage{
aoristic_vm_mle(d, kp_max = 100, ...)
}
\arguments{
\item{d}{Data frame with columns \code{t_start} and \code{t_end}, which are
the start and end times.}

\item{kp_max}{Maximum value of \code{kp} to search for.}

\item{...}{Additional arguments for the \code{optimize}.}
}
\value{
Numeric vector of length two; estimates for mean direction \code{mu}
  and concentration parameter \code{kp}, respectively.
}
\description{
Get maximum likelihood estimates for the von Mises distribution based on
aoristic data.
}
\examples{
dat <- generateAoristicData(n = 200)
aoristic_vm_mle(dat)
}
