% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{aoristicVM_mle}
\alias{aoristicVM_mle}
\title{Get maximum likelihood estimates for the von Mises distribution based on aoristic data.}
\usage{
aoristicVM_mle(d, startingValues = c(0, 1))
}
\arguments{
\item{d}{Data frame with columns `t_start` and `t_end`, which are the start and end times.}
}
\value{
Numeric vector of length two; estimates for mean direction `mu` and concentration parameter `kp`, respectively.
}
\description{
Get maximum likelihood estimates for the von Mises distribution based on aoristic data.
}
\examples{
dat <- generateAoristicData()
aoristicVMMLE(dat)
}
