% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularUtilities.R
\name{dcunif}
\alias{dcunif}
\title{Circular Uniform Distribution}
\usage{
dcunif(x, from = 0, to = 2 * pi)
}
\arguments{
\item{x}{Angle to evaluate.}

\item{from}{Lower bound of the circular uniform distribution.}

\item{to}{Upper bound of the circular uniform distribution.}
}
\value{
Numeric; probability of x.
}
\description{
Evaluate the circular uniform distribution.
}
\examples{
dcunif(1, 6, 2)

}
