% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AoristicFunctions.R
\name{generateAoristicData}
\alias{generateAoristicData}
\title{Generate aoristic data}
\usage{
generateAoristicData(n = 30, trueDistGen = function(n)
  as.numeric(suppressWarnings(circular::rvonmises(n, 1, 10))),
  intervalSampler = function(n) runif(n, 0, 2),
  LBSampler = intervalSampler, UBSampler = intervalSampler,
  aoristicProportion = 1)
}
\arguments{
\item{n}{The desired sample size.}

\item{trueDistGen}{A function that takes a single argument n and samples n
random values from the desired true distribution.}

\item{intervalSampler}{A function that takes a single argument n and samples
n random values that determine how far away the interval bounds are from
the true sampled datapoint on either side. By default the same function is
used on both sides, but this can be overwritten by specifying LBSampler and
UBSampler.}

\item{LBSampler}{Optional function of the same form as intervalSampler that
specifies the way in which the distance between the lower bound and the
true value is sampled. Should sample positive values which represent the
difference.}

\item{UBSampler}{Optional function of the same form as intervalSampler that
specifies the way in which the distance between the upper bound and the
true value is sampled.Should sample positive values which represent the
difference.}
}
\value{
A data frame containing the start, end, and true values.
}
\description{
This function generates aoristic circular data from a true distribution
generation function and two function to generate the lower and upper
distances from the true point.
}
\examples{
generateAoristicData()

}
