% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AoristicFunctions.R
\name{getCircAoristicFunction}
\alias{getCircAoristicFunction}
\title{Aoristic Fraction function}
\usage{
getCircAoristicFunction(df)
}
\arguments{
\item{df}{A `data.frame` or matrix of which the first column represents the
lower bound and the second column represents the upper bound of the
observed intervals.}
}
\value{
The aoristic function.
}
\description{
Create a function for circular interval censored analysis. The created
function takes a vector `x` and returns the height of the aoristic fraction
given that function.
}
\examples{
df <- generateAoristicData(n = 5)
myfun <- getCircAoristicFunction(df)
ggplot(data.frame(x = c(0, 2*pi)), aes(x)) +
  geom_hline(yintercept = 0, color = "gray") +
  stat_function(fun = myfun, n = 1000) +
  coord_polar() +
  ylim(-1, NA) +
  theme_void()
}
