% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularUtilities.R
\name{meanDirAoristic}
\alias{meanDirAoristic}
\title{Mean direction of aoristic data.}
\usage{
meanDirAoristic(d, na.rm = TRUE)
}
\arguments{
\item{d}{Aoristic data frame.}

\item{na.rm}{Logical; If \code{FALSE}, returns NA if there are missing
values.}
}
\value{
The mean direction in radians.
}
\description{
Compute the mean direction of an aoristic data frame. It is the mean
direction of the midpoints.
}
\examples{
meanDirAoristic(generateAoristicData())

}
